# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BLASTERR social platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{username}:
    get:
      operationId: getUserProfile
      tags: [people]
      summary: Get a user profile
      parameters:
        - $ref: "#/components/parameters/Username"
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: User not found
...
  /reports:
    post:
      operationId: createReport
      tags: [blasts]
      summary: Report content or a user
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "401":
          description: Authentication required
...
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: List users for moderation
      security: [{ clerkSession: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [all, active, suspended] }
      responses:
        "200":
          description: Admin users
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUsersResponse" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update a user's moderation status or role
      security: [{ clerkSession: [] }]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminUserUpdate" }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUser" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
        "404": { description: User not found }
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user
      security: [{ clerkSession: [] }]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204": { description: User deleted }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
        "404": { description: User not found }
  /admin/content:
    get:
      operationId: getAdminContent
      tags: [admin]
      summary: List Blasts for moderation
      security: [{ clerkSession: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [all, published, hidden, removed] }
      responses:
        "200":
          description: Moderation content
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminContentResponse" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/content/{id}:
    patch:
      operationId: updateAdminContent
      tags: [admin]
      summary: Update a Blast moderation status
      security: [{ clerkSession: [] }]
      parameters:
        - $ref: "#/components/parameters/BlastId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminContentUpdate" }
      responses:
        "200":
          description: Updated Blast
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminContentItem" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
        "404": { description: Blast not found }
    delete:
      operationId: deleteAdminContent
      tags: [admin]
      summary: Remove a Blast
      security: [{ clerkSession: [] }]
      parameters:
        - $ref: "#/components/parameters/BlastId"
      responses:
        "204": { description: Blast removed }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
        "404": { description: Blast not found }
  /admin/reports:
    get:
      operationId: getAdminReports
      tags: [admin]
      summary: List community reports
      security: [{ clerkSession: [] }]
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [all, open, in_review, resolved, dismissed] }
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminReportsResponse" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/reports/{id}:
    patch:
      operationId: updateAdminReport
      tags: [admin]
      summary: Resolve or triage a community report
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminReportUpdate" }
      responses:
        "200":
          description: Updated report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminReport" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
        "404": { description: Report not found }
  /admin/moderation:
    get:
      operationId: getAdminModeration
      tags: [admin]
      summary: Get the moderation queue
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Moderation queue
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminModerationResponse" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
    post:
      operationId: createAdminModerationAction
      tags: [admin]
      summary: Take a moderation action
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminModerationAction" }
      responses:
        "200":
          description: Moderation action recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminActionResult" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
        "404": { description: Report not found }
  /admin/analytics:
    get:
      operationId: getAdminAnalytics
      tags: [admin]
      summary: Get platform analytics
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Analytics
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAnalytics" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/system:
    get:
      operationId: getAdminSystem
      tags: [admin]
      summary: Get protected system health details
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: System health
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSystem" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/audit-log:
    get:
      operationId: getAdminAuditLog
      tags: [admin]
      summary: List moderation audit events
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Audit log
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAuditResponse" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get moderation settings
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSettings" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
    patch:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update moderation settings
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminSettingsUpdate" }
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminSettings" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/announcements:
    get:
      operationId: getAdminAnnouncements
      tags: [admin]
      summary: List platform announcements
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Announcements
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAnnouncementsResponse" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
    post:
      operationId: createAdminAnnouncement
      tags: [admin]
      summary: Create an announcement
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminAnnouncementInput" }
      responses:
        "201":
          description: Created announcement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAnnouncement" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/announcements/{id}:
    patch:
      operationId: updateAdminAnnouncement
      tags: [admin]
      summary: Update an announcement
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminAnnouncementUpdate" }
      responses:
        "200":
          description: Updated announcement
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminAnnouncement" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
        "404": { description: Announcement not found }
    delete:
      operationId: deleteAdminAnnouncement
      tags: [admin]
      summary: Delete an announcement
      security: [{ clerkSession: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Announcement deleted }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
        "404": { description: Announcement not found }
  /admin/features:
    get:
      operationId: getAdminFeatures
      tags: [admin]
      summary: List feature controls
      security: [{ clerkSession: [] }]
      responses:
        "200":
          description: Feature controls
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminFeaturesResponse" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
  /admin/features/{key}:
    patch:
      operationId: updateAdminFeature
      tags: [admin]
      summary: Toggle a feature control
      security: [{ clerkSession: [] }]
      parameters:
        - name: key
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdminFeatureUpdate" }
      responses:
        "200":
          description: Updated feature
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminFeatureFlag" }
        "401": { description: Authentication required }
        "403": { description: Admin access required }
        "404": { description: Feature not found }
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
  parameters:
...
    BlastId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    Username:
      name: username
      in: path
      required: true
      schema:
        type: string
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
...
    Profile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [blasts, media]
          properties:
            coverUrl: { type: string }
            blasts:
              type: array
              items: { $ref: "#/components/schemas/Blast" }
            media:
              type: array
              items: { $ref: "#/components/schemas/Blast" }
...
    ReportInput:
      type: object
      required: [targetType, targetId, reason]
      properties:
        targetType:
          type: string
          enum: [blast, comment, user]
        targetId: { type: string }
        reason:
          type: string
          enum: [spam, harassment, threats, hate, nudity, violence, scam, personal-information, impersonation, other]
        description: { type: string, maxLength: 1000 }
    Report:
      type: object
      required: [id, status]
      properties:
        id: { type: string }
        status: { type: string }
...
    AdminUser:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required: [role, status]
          properties:
            role: { type: string, enum: [user, moderator, admin] }
            status: { type: string, enum: [active, suspended] }
    AdminUsersResponse:
      type: object
      required: [users, total]
      properties:
        users: { type: array, items: { $ref: "#/components/schemas/AdminUser" } }
        total: { type: number }
    AdminUserUpdate:
      type: object
      properties:
        role: { type: string, enum: [user, moderator, admin] }
        status: { type: string, enum: [active, suspended] }
...
    AdminContentResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/AdminContentItem" } }
        total: { type: number }
...
    AdminAnnouncementsResponse:
      type: array
      items: { $ref: "#/components/schemas/AdminAnnouncement" }
...